pyproject.toml
[project]
name = "da-radi-boga-mnie-nie-zhalko"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "matplotlib>=3.10.8",
    "python-dotenv>=1.2.2",
]
